Remove delegated current_edition methods from document
Using these didn't seem to add much value and seems to introduce some
confusion. Using a current_edition_title vs current_edition.title
doesn't save any characters.

So the main reason for these to exist is to conform to the Law of
Demeter, which we can always go back to if needed.