Change all foreign keys to restrict on deletion
Previously we had a variety of on_delete actions for our foreign keys,
in an effort to try and facilitate deleting stuff, but without making it
too easy. As the logic for these actions was somewhat complex and we
don't actually have a use-case for deleting anything, ever, it makes
sense to just restrict deletion of anything that's referenced somewhere.