Block changing revision on a live edition
This is a somewhat crude exception that prevents the situation where a
user is able to edit a live edition by preventing this with an exception
at the point of assigning a revision.

I'm writing the fateful words that haunt a developer in later years that
this is intended to be temporary. Ideally we'd have a view or an alert
presented to the user to describe the edition is now live and you should
create a new edition to make changes.

Before versioning was introduced this behaviour was managed by the
PreviewService which would automatically create a new edition if a user
managed to edit a live edition. This behaviour doesn't feel correct as
creating a new edition is quite a big step for a document so a user
shouldn't be unaware that they have just done that (unless we were to
focus the UI on making this more obvious).

With this an attempt to do this will cause our application to raise an
error and give the user an error screen, which is not ideal but is
simple and doesn't risk inconsistent data.

The expectation is for this to stand until we can define a
service/process that can deal with conflicts like this and cases where a
user is editing a stale item.