Raise exceptions when user requests fail pre-conditions
This replaces two instances of redirects (one of which forgot a
return 😬) for cases where the edition pre-conditions failed which stops
the request. In more plain terms, you can't create a new edition if you
have a draft (as that would have 2 drafts) and you can't review
something that doesn't need review.

These shouldn't really be exceptions but right now we don't know exactly
what they should be and how we present it. We have a bunch of these in
the app (e.g. editing a live edition in many places) so it'd be good to
resolve these together.