Rename ambiguous revision method names
We had two pairs of confusing method names on Revision: #status
and #statuses; and #revision and #revisions.

This attempts to make these more distinct by prefixing the many variants
with all. So we now have #all_statuses and #all_revisions.

The other option on this was to prefix the singular versions e.g. if we
had #current_status and #current_revision. However this feels
unnecessarily verbose since revision and status will be frequently
accessed whereas the collection equivalents are used for audit trails
and will be rarely used.

I'm not 100% sure the `all_` prefix is the best one, I thought a better
one for statuses would be #statuses_held but that didn't seem to fit
nicely with revisions.