Remove old assets at the point of previewing
This uses the point of creating a preview as a point to delete assets
from Asset Manager which are no longer used.

When an edition is sent to the preview service the previous revision
assets are compared with the current revision and any that are in a
draft state are then deleted from Asset Manager.

This doesn't feel like the most robust of solutions as if assets are not
removed at this point they will remain in draft forever. The approach
that would resolve that would be to have some form of scheduled task
that performs a cleanup of assets periodically. However since we don't
have prior art currently in scheduled tasks on this app I have not
implemented it.