Remove statuses and revisions many associations on Edition
This removes these associations as they are not currently used. Their
existence seems to have been somewhat divisive due to us having a status
and statuses, and revision and revisions.

There is a counter proposal to renaming status and revision to
current_status and current_revision. This is not the preferred course of
action as it leads to a somewhat superfluous spread of the word
"current" through the codebase.