Test Asset Manager receives lead image
This enhances the upload_lead_image_spec to also check that files are
uploaded to Asset Manager.

This is to fill a gap that we didn't have feature tests that tested
whether assets made it to Asset Manager. This test was chosen as it was
the point images would be uploaded to Asset Manager for the first time.