Widening the initialiser for GovspeakDocument
This change required making changes that ensured the controllers that called
GovspeakDocument would have access to document_id - an attribute that can be
used to obtain the current edition for a document that is now expected in
the class's constructor.

Many of the changes relate to threading a reference to document_id through
various levels of view code and components so that the preview controller can
make use of it.