Reorder unpublish service withdraw method
This reorders the withdraw method so that an edition is only unpublished in the
Publishing API after successful writes are made to the database.  Normally, if
the Publishing API call fails the transaction would roll back and the database
would not be updated.  However as withdraw contains a transaction sitting
within a transaction (in the Withdraw controller) then when such an exception
is raised a rollback is not created. By setting `requires_new` we create a save
point in our transaction and any unhandled exceptions in this block will be
rolled back.

Relevant part of rails docs:
https://api.rubyonrails.org/classes/ActiveRecord/Transactions/ClassMethods.html#module-ActiveRecord::Transactions::ClassMethods-label-Nested+transactions