Add Factorybot linting to rake
This adds a linting rake task for our factories. It runs automatically
when running rake as it's been added to rakes default tasks. You can
also run it manually by simply running the specific rake task (E.g rake
factorybot:lint).

The lint task is contained within a transaction with a `raise
ActiveRecord::Rollback` which ensures all factories created in the
linting process are rolled back to leave a clear db.

Also, i've removed the `spec` arg to the default tasks as rspec-rails
already adds this to the top of the default task list ran by rake.