Run factorybot linter after db has been set up
This was causing test failures due to the linter being ran against an
old version of the db. As rails db:reset had not been run yet and there
was still migrations pending the linter was trying to create factories
against old tables (column names) which caused the lint to fail.

Moving this to run after the tests means the db is set up correctly for
the linter.