Update autocompleteWithHintOnOptions to use select element
Previously this was removing the select from the DOM which meant that
there was actually no form field being submit. This was causing the add
contact feature to fail as it was treating all submission as empty
input. This wasn't caught by our test suite as we were testing without
JS.

This updates the script to stop detaching the select element from the
DOM and instead hide it. It also adds the value attribute of the option
to the JSON passed to autocomplete and that is used to match an option.
This was done mostly as it felt a bit brittle to be matching text but I
now realise it gives us the potential to distinguish between two contacts
with different organisations but same name.

Since we now don't hide the select element we no longer need to create
the .autocomplete__wrapper, which is nice since that meant we ended up
with two .autocomplete__wrapper elements.