Remove custom_template option
I had thought we'd removed this terminology in the initial PR for this
so this seems to be an oversight.

Rather than replacing it with a boolean attribute I felt this made more
sense to be something that could be specified that could only be one,
since if we have a different variation of this we'd not want to allow
both to share a boolean. For this I decided to go with module as that
affects the data-module attribute that's set. Since module is a Ruby
reserved word I then switched to data_module as that was preferred to
js_module. But well I don't think I stumbled upon a winner here
so other ideas welcome.