Support Images for Govspeak Documents
The focus of this commit is on supporting the different needs that Preview and
payload contexts have for accepting HTML code containing images.  to_html has
been turned into separate methods in_app_html and payload_html, which differ only
in the kind of value that is passed as a URL attribute for when Govspeak generates
an HTML image snippet.

Payload uses a URL for an image that comes from asset manager, whereas the preview
feature uses the path for a cropped variant of the image. Initially,
image_revision.crop_variant.service_url provided the URL attribute for previews.
However service_url requires that a host is specified, and this is not provided
within the service tests.  Using the rails_representation_path helper method avoids
this problem.

Code used to support in_app_html and payload_html showed a lot of similarities, so
some of the changes have helped minimise repetitive code.  Private methods in_app_images
and payload_images are meant to help make it obvious that the orchestration methods
in_app_html and payload_html do different things.

Splitting to_html into in_app_html and payload_html in govspeak_document warrants
changes to be made to the tests in govspeak_document_spec. Separate describe blocks appear
for each method.  They currently exhibit a lot of repetition that may be remedied in a
subsequent commit.