Accessible autocomplete test helper
This adds a helper method that can be used in feature tests that involve
the accessible autocomplete component.

We have to initially wait for the autocomplete to initialise, this
involves waiting for an input element with the id we expect, then we can
fill this in and select the item we want. We then  need to wait for JS to
select an option element before we are safe to submit the form.