Make UnpublishService save the published_state of the doc
UnpublishService now adds the published_state of the
document to the withdrawal record when withdraw is called.

This allows us to persist the published_status before the document was
withdrawn which enables us to revert back to this status if needed, even
after it has been withdrawn multiple times (e.g changing the explanatory
note).