Override markdown preview style for inline images
Previously the margin on the figure element was causing the markdown
preview for an inline image to extend beyond the width of the preview
dialog. This adds a scoped override to remove this unwanted margin,
although in future we should fix this to make use of the styles for the
frontend apps that achieve the same thing.