Stop changing text of metadata "Save" button
Previously the text of this button would change depending on whether the
user was uploading a (lead) image. This implicit behaviour is now made
explicit with the checkbox, so the button text doesn't need to explain.

This also replaces the vague and potentially brittle uses of the 'Back'
button as a way of getting back to the document summary. It also removes
an unnecessary check for not sending an image payload when editing meta.