Add initial checkbox and controller logic
This adds a new checkbox to the metadata form for images, which can be
used to select/remove a lead image at the same time as changing the
metadata for that image. As the existing controller action is quite
complex, this commit focusses on making the smallest change to add this
new functionality, with the aim of refactoring it in a future commit.