Replace revision build_ methods with services
Previously our revision models had a build_revision_update method to
handle creating a new revision tree when certain attributes change.

Having the versioning logic directly on the models makes it cumbersome
to figure out which attributes are changing; in particular, it's not
compatible with the changed_attributes feature provided by ActiveRecord.
This makes it hard to answer questions like 'has the lead image
changed?', which is necessary in order to correctly redirect the user.

This extracts the versioning logic into services, which provide changed?
and changed_attributes methods to better support our controllers. I've
used inheritance to reduce the amount of duplicated versioning logic for
sub-revisions like Image::FileRevision and Image::MetadataRevision.

    updater = BaseUpdater.new(<revision>, <user>)
      .track_attributes(%i[title summary blah])

    next_revision = updater.assign_attributes(title: "new title")

    do_something if updater.changed?
    do_something_else if updater.changed_attributes.key?(:title)

Some of the actions on (lead) images require custom versioning logic to
figure out what the new values of attributes should be. It seems best
to have this logic in the ImagesController so that both it and the
controller can make use of the updater methods. Hopefully this will be
refactored as part of a larger overhaul of the controllers.