Conditionally change the Withdraw submit button
This changes the submit button for the withdraw form to be different
whether or not an edition is already withdrawn.

This switches both the methods in this controller to use @edition as an
instance variable rather than @document and updates the views accordingly.

It also aims to make the feature test more feature focused.