Make brakeman happy
It didn't like us redirecting directly to a model that was then resolved
as a URL. Presumably because there's a param input involved. This seems
to be a case of brakeman being a bit over zealous but since the change
makes the route more explicit it doesn't feel like a bad thing.