Resolve issue where withdrawal time changes
We previously used the time a status was created to show a user when a
piece of content was withdrawn, however this is not accurate if you
update the explanation of a withdrawal as this creates a new status.

Thus the way I've approached resolving this is by passing a timestamp
among the details of Withdrawal objects so that this data is persisted.
If a user undoes a withdraw and re-withdraws this timestamp will be
reset.