Refactor update action for images controller
Previously the decision logic for building a new revision, determining
the timeline entry and doing the redirect was quite complex. This is an
attempt to simplify the logic by:

   * Extracting the boilerplate to create a TimelineEntry
   * Adding clearer selected?/removed? question methods
   * Adding a clearer next_lead_image_revision procedure
   * Adding an image/lead upsert method to revision

Because a new revision can't be modified once it's built, we need the
a specific method on the Revision class to avoid having to having to
manually determine if a new one needs to be built.