Move building a withdrawal to a separate method
This was triggering the linter to complain about the size of the block
of code, which I've take as a good sign that this should have some of
it's code reduced.

Since this seems to be the obviously logic heavy part of the code it
seemed the best candidate for separating out.

I think in time we'll want to reduce this separate this service into
distinct ones for withdrawing and removing.