Split UnpublishService into WithdrawService and RemoveService
This follows up on
https://github.com/alphagov/content-publisher/pull/757 by splitting the
UnpublishService into two distinct classes.

This was done as the approach to withdrawing compared to removing was
diverging increasingly to the point that there is low common
functionality or procedure.

With the RemoveService I have combined what was previously two methods
(remove and remove_and_redirect) into a single call method and instead
passed in a Removal object as an argument. This is inconsistent with how
the WithdrawService works but these have somewhat diverged and I'd
expect once we allow removing from the UI that this will need amends.

Due to the change in interface for RemoveService it wasn't trivial to
maintain the same testing approach (checking that they call a service)
so I've decided to do these as integration tests. This also fills a gap
that there wasn't an integration test that tested these tasks do what is
expected.