Remove redundant uses of 'wizard'
Previously we used a 'wizard' query param to determine the next/previous
page when editing image crop/metadata. This is no longer necessary, as
the next/previous pages are always the same.

In order to continue showing a custom flash message when a lead image is
deleted, this replaces the use of 'wizard' with a specific check that
the lead image has been removed (as part of the deletion).