Fix visited state on link to change withdrawal
Previously the 'notice' component would accept govspeak as it's block
input, but this has changed to HTML in the most recent version, such
that the applied class now works. This manually renders the withdrawal
explanation with Govspeak to preserve the desired behaviour, as well as
replacing the <p> elements with a <br> to preserve the previous spacing.