Fix and extend modal tests to cover interface
Previously the modal tests were calling .focus() directly, which doesn't
work in headless mode; this fixes the tests to manually dispatch the focus
event. One test has been removed, which used document.activeElement to
return focus to an element when the modal closes (broken in headless).

This also rewrites the tests to be more focussed on the interface
provided by the modal component, similar to our other components, while
also keeping some of the end-to-end tests to ensure all is covered.