Replace internal modal title with a title ID
Previously the modal had a title option and populate its own title
element that would act to describe the modal, using aria-labelledby.
Using static markup for the title makes it difficult to show existing
pages in the modal, which already contain markup for their own titles.
This removes the static title markup within the modal component and
instead adds an mandatory option to specify which title element to use.