Remove aria-labelledby attribute on modal
Previously we used the attribute to identify the function of the modal,
but testing with a screen reader shows this is redundant, as the reader
simply reads the title twice. This removes the redundant attribute, as
we believe it's sufficient that the modal content provide a heading that
describes what the modal is about; no further markup seems necessary.