Add component to show one section at a time
This component is intended to show a single block of HTML (a 'section').
Sections can be specified in the config, or dyanmic content can be shown
in a special section, using the API exposed by the component's module.

This component has been created to support the modal work on inline
images, which involves replacing the entire content of the modal
multiple times, either with a static screen, such as a loading screen,
or with dynamic content that forms part of the inline image workflow.

In order to avoid any flickering, the content of each section is wrapped
with a class to hide it by default, using styles to override this. Since
the point of this component is to only show a single section, if JS is
not available, the component shouldn't be used and shows nothing.

Because the content of this component can change, we set
aria-live="assertive" in order to notify any assistive tech, much like
when a browser page reloads. The downside is that any section with a
'noisy' component will require work to ensure it doesn't annoy the user.