Make existing image partials context-sensitive
This sets a @context variable for the images index action, in order to
render the images index page in a way that's appropriate for display in
a modal. Making the page context-sensitive means we can re-use the
existing pages, instead of maintaining separate ones for the modal.

We set the @context variable based on a standard header, which means we
can utilise the Rails .xhr? method to check it's set, and avoid exposing
a query parameter in the URL. In future work, we will need to look at
extending the modal layout to support error messages and 'Back' links.