Add inline image modal for markdown editor
This adds a JS 'glue' module called 'inline-image-modal' that modifies
the new 'Insert image' button on the markdown editor, so that it opens
the modal with the modal version of the images index page. We then also
override the 'Insert image markdown' action to insert into the editor.

It's unclear how much of the new 'inline-image-modal' module is generic,
but there seem to be a few key functions that we will want to extract:

   * fetchModalContent performs an authenticated get request and request
and returns some HTML in a promise
   * overrideActions scans for 'data-toggle-action="<name>"'s and
overrides click events to call performAction(<name>)
   * performAction takes an action like 'open' or 'insert' and does a
lookup to see which callback to invoke

Writing the 'glue' module in this way makes it easy to see the available
actions, and separate them from generic code like overriding buttons.