Send user uid value to Google Tag Manager
This is so we can monitor session behaviour of particular users and see
when they're using different browsers/devices.

According to Andy it is acceptable privacy wise to send an id that has a
meaning in our internal systems but cannot be used out of context to
identify the user. I debated whether it would be better to run this
through a one way cryptographic hash function such as SHA1 but I'm not
sure that offers any actual benefits.