Simulate initialisation of error-summary
This manually focusses any error-summary component to simulate its
normal behaviour. Ideally, we should change this and other components in
govuk-frontend so that they can be re-initialized, and in this case,
stop relying on window events to auto-focus.

We also have components from govuk_publishing_components, which can be
idempotently re-initialized with 'GOVUK.modules.start()', but this is
not necessary yet. And we have our local components, which don't support
re-initialization at all, which we will address in future work.