Support cropping images in modal
This changes the upload workflow in the modal, so that we now show the
crop page after the image is successfully uploaded; saving the crop then
takes the user back to the images index page, showing the cropped image.

Related to this, the modal version of the images index page now includes
the 'Edit image' action, which also takes the user to the crop page.