Modify when change note guidance appears
This passes a data_attribute for contextual guidance to the radio button
component, this means that change note guidance will only appear when a
user selects a major change and will disappear if the user selects a
minor change.

JS for contextual guidance adds event listeners to all elements with
data attribute of `data-contextual-guidance` to handle displaying
and hiding a components guidance, this enables the minor / major change
note radio button to benefit from this functionality, instead of before
where only focusing on the textarea would display the guidance.

Dependant on this for govuk_publishing_components:
https://github.com/alphagov/govuk_publishing_components/pull/766

Trello:
https://trello.com/c/pcbHvl7B/636-change-note-guidance