Save pre-scheduled status and review status
In the event that a scheduled publishing is cancelled, we need to know which
state the document was in so that we revert back to it. We also need to know
the review status of the document so that we know if it has been effectively
"force scheduled" and therefore let users know that it needs to be reviewed
after it has been published.