Support editing image metadata within a modal
This changes the post-crop behaviour to now show the metadata edit page
for the cropped image, which now matches the upload/edit image workflow.

This also rewrites the feature test for uploading an inline image to
more closely ressemble the non-modal equivalent. As it's not possible to
manually navigate to a modal section, the metadata requirements feature
spec has to manually navigate to the page via the 'Edit image' workflow.