Support 'Save and insert' on inline image upload
Previously we redirected back to the image index screen irrespective of
whether the user was uploading a new image or editing an existing one.
This (re-)introduces a 'wizard' parameter to remember that the user was
uploading an image, and switches the save action to 'Save and insert'.

Becuase saving image metadata can fail due to requirements issues, this
also modifies the helper functions in the inline-images module to return
a richer response object that indicates whether the post action was
'successful'. For now, it's enough to check that a redirect occurred.