Override form submit action instead of button
Previously we overrode the behaviour of the 'Upload' button to support
the inline image workflow, but this interferes with our GTM 'submit'
listener. This changes the way we override the 'Upload' action, so that
it happens on the form instead of the button - with some changes to
our GTMFormListener module to support re-initialisation, this works.