Don't display schedule action if content is not schedulable
If an edition has a set scheduled publishing datetime that is in the past, we
don't display a "Schedule" button. One possible edge case here is a user could
be on the summary page while the set datetime becomes past and the "Schedule"
button would still appear, so we also raise an exception in the Schedule
controller in case this should occur.