Support deletion of images in a modal context
This adds a delete action for all non-lead images in the inline-images
modal. As we only show the lead image for context, this action is not
supported for it. One slight surprise was that we actually do a POST
request for deletions, using a hidden parameter to override the method.