Fix accidental submit when choosing images
Previously we moved some of the override behaviour of the inline image
modal to apply to the 'submit' form, instead of the 'click' event of
the submit button of the form.  However, the 'click' event was still
being handled on the form, which was interfering with its other inputs.

This separates the events we override for modal content: 'submit' for
forms and 'click' for links. It also removes the (now) redundant data
attribute to specify the ID of the form element, as this is now the same
as the action item (element) being performed.