Merge autocomplete and its hinted variant
Previously we had a flag that would change the autocomplete to add hint
text to each entry. The hinted variant is only used for contacts, and
was broken due to a misplaced data attribute.

This merges the hinted variant with the normal non-hinted autocomplete,
which fixes the issue with contacts and removes some duplicated code,
noting that the hinted variant also supports not using a hint.

This also removes support for contextual guidance, since this is not
currently being used with the autocomplete, and in any case was only
implemented for the non-hinted variant.