Begin replacing model build methods with updaters
Previously we used a tree of 'build_revision_update' methods on each
model, which has lead to some duplication between models and quite a lot
of extra code in the model classes. This makes a start at extracting
this logic into a set of 'updater' classes with some helper methods.

The updaters take a hash of 'fields', which can include columns,
collections and sub-fields for sub-revisions. I've used inheritance to
DRY-up some of this logic, as well as create a clear interface for the
updater classes that derive from it.

Some of the existing code doesn't fit very well into the updater
classes, as it's specific to a single use-case and doesn't make sense to
share centrally. This does mean the ImagesController now has a bit more
code in it than before. A nice compromise that would avoid bloating the
existing controller and the new RevisionUpdater, would be to use an
'include' pattern to add more 'assign' functions to RevisionUpdater.