Extract initComponent/overrideActions into utility
Neither of these functions are not specific to the inline image workflow
and it makes sense to move them to an external class to enable re-use.

Although it's tempting to move some of the other functions, such as the
ones to support rendering, it's not yet clear how coupled these are to
the specific needs of the inline-image workflow, including its use of
the multi-section-viewer; therefore, it seems best to leave them alone.