Configure byebug for procfile running
With the addition of sidekiq workers the expectation is that this
application will be ran via the procfile in development environments.

This poses a challenge for using byebug as it does not usable within a
foreman output.

To remedy this I've set up remote debugging with byebug and assigned
default ports to be used, these ports should be reserved in the GOV.UK
development VM.