Redirect /documents to include organisation by default
When a user accesses /documents and has no query string parameters this
redirects the user to
/documents?organisation=#{current_user.organisation_content_id}

This is so that a user has their organisation as the default filtering
mechanism and that this filtering is done by a URL that can be shared no
matter which organisation a user is in.

This feature is currently behind the pre-release features permission
until it is released to users.