Re-arrange the index filters
This follows a discussion with Sonia where it was suggested that it'd be
easier to navigate these filters if organisation was showed near the top
to make it clearer that this is pre-selected.

Then document type was moved to be below status as it contains only
small amounts of information now and is unlikely to be used.

Finally the clear all filters now shows all organisations. This is a
change from my original implementation of this where it reset the form
to the users default view. I'm not surely what is more intuitive, reset
feels like the correct behaviour but the "clear" wording does not imply
reset.

Note, this is not behind a feature flag as it is not part of the new
feature.