Fix tests broken by index organisation filter change
These tests needed amends so that they would continue to pass with the
change in index page behaviour to redirecting when an organisation is
applied.

These conceal the index change somewhat to try decouple these from it.
This seems to be a pro for decoupling and something of a negative for
transparency.