Move image update code into updater module
Previously the ImagesController contained a lot of code to work out the
next set of image revisions and the next lead image revision as part of
the user making a change. This moves the code into a new module that
acts as an extension to the RevisionUpdater and refactors it a bit.