GTM listener for selected topics
This sets up an event handler for window.submit events that are to send
the topics a user selected to Google Tag Manager.

It listens on forms tagged with a data attribute of gtm-selected-topics
and uses the API on the custom miller-columns-selected element to
retrieve the list of selected topic names.