Return a 422 status code when a request is unprocessable
For cases where a user submits a form and we return it back with errors
(or issues in our terminology) return a 422 status rather than a 200
one.

This change makes it easier for us to fix a bug where it's hard to
determine whether a request was successful or not as part of the modal
flow.

https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/422 for more on
the status code.