Monitor ModalFetch responses for 422 status codes
This changes the approach of the fetch handling due to an issue with the
fetch polyfill. It does not support response.redirected [1] so another
approach was needed to support browsers without native fetch ☹️.

I felt the simplest way was to inverse this from needing a done flag to
instead track whenever an unprocessable entity is returned and flag
that. Then we can assume that the lack of this flag means a successful
request.

[1]: https://github.com/github/fetch/blob/794f6df048922427398c26957eb6ae40fd649a51/fetch.js.flow#L52-L57