Make the generic GtmFormListener into something specific
This makes a number of changes to the GtmFormListener module to remove
problems with it.

Firstly, there are problems accessing the data that is sent to GTM as it
is not sent in the normal GTM event object. Instead it used a custom
event of the form: `{ eventName: { eventKey: eventValue }`. This has now
been changed to send data in the form of `{ event: eventName, value:
eventValue }` which matches GTM conventions.

The event data we have to pass through is the name of the form, the
field they selected and the value. The quickest way I was able to do
this was combining this data into two fields: event name and value and
this is done in this form:

```
{
  event: "checked-inputs.${formName}",
  value: "${fieldName}: ${fieldValue}"
}
```

I chatted to Andy about this and he liked this structure, so this is
staying rather than changing into something with more use of fields on
the event object.