Use a trait to share revision fields
This is to reduce repetition between these two files (as I'm just about
to make this worst).

This isn't without drawbacks though, it definitely is less clear what
all the transient attributes of an edition is while using this. On the
other hand it does save repeating these defaults in multiple places.