Set up AuthenticationHelper for feature tests
This allows specifying the user that is logged into a feature test by
setting the GDS::SSO.test_user. If there isn't a test user set we can
assume the current user is the first user in the database - as that's
how GDS::SSO works.

This allows us to stop referencing `User.first` in tests, which requires
domain knowledge to understand that it is the logged in user. It also
lets us login as different users, rather than modifying the default
user.

This doesn't have a logout method since this application doesn't support
a user unauthenticated - and I'm not quite sure how to do that with
GDS::SSO simply.

A caveat of this is that once logged in it doesn't seem to check again
in the process of a test. So you can only run login_user before tests
begin. There are probably ways of avoiding this by interacting directly
with warden, but this doesn't seem necessary now.