Edition methods to load current edition with joins
This is intended to replace the need to have the following methods in
controllers:

```
document = Document.with_current_edition.find_by_param(params[:id])
edition = document.current_edition
```

and

```
Document.transaction do
  document = Document.with_current_edition.lock.find_by_param(params[:id])
  edition = document.current_edition
end
```

with:

```
edition = Edition.find_current(document: params[:document])
```

and

```
Edition.find_and_lock_current(document: params[:document]) do |edition|
end
```

This is intended so that controller code can be more focused on an
edition, as most operations require a current edition.

The scope used for this accepts a hash where a current edition can be
looked up via either an id key or a document key. The id key is an
integer of the edition id. The document key is expected to be a document
param - a string containing content_id and locale. Were needs to change
for more document related lookup this could query differently depending
on the input, e.g. it could lookup for a document if given that or a
param string.

The aspect of this I'm least clear on is what the best naming for these
methods should be. I've gone for succinct, low detail names that hopefully
reflect a users intention.