Create abort middleware for Sidekiq
This sets up a middleware for sidekiq to provide the means to abort from
a worker and have this not cause any automatic retries.

The reasons for adding this are:

- To allow means to explicitly abort a worker from any point in a
worker, as opposed to only a return call in the perform method;
- To be able to provide direct feedback on worker aborting in test and
console scenarios.

Adding this is pretty simple but does mean some annoying decisions about
where to place code. The middleware itself seemed to belong in a lib
directory I went for /lib rather than creating /app/lib since that does
not exist in this app.

The exception also needs a place to live too, I decided against following the
[errors dir][1] as used in Publishing API and felt that putting it with
the middleware was the most logical place. It felt a bit wrong that it
wasn't under a module namespace, but that felt like it'd add unnecessary
verbosity to the code.

[1]: https://github.com/alphagov/publishing-api/tree/979479f908a06fe592cd6818c42e2cd6ca064eef/app/errors