Schedule the Scheduled Publishing Worker
We schedule the Scheduled Publishing Worker to perform a job at the time that
content has been scheduled to publish.

By default Sidekiq warns when jobs produced by tests aren't processed. We don't
want test jobs to be processed and the warning is a bit noisy so we configure
`rspec-sidekiq` to suppress it.