Create unschedule action
This adds a new controller action to handle when a user wants to cancel
a scheduling. This creates a new timeline entry with a type of
`unscheduled` and assigns the document its previous state before it was
scheduled (draft, submitted_for_2i etc).

The datetime is still set on the document so the user can easily
schedule to the same time again.