Check govspeak body fields are valid
This adds govspeak validation, which is similar to what is in Whitehall [1]
and Specialist Publisher [2]. It is added for the body govspeak field -
in theory it applies to all govspeak fields but there is only error
messages for body.

Writing the validation message is rather grim as there are quite a few
reasons the `Govspeak#valid?` method can fail. I felt the simplest way
to resolve this was just to tell users to "Remove HTML" - this hasn't
been content designed though and may need revising if there are
situations where users are actively advised to enter HTML.

Finally to perform the valid method we need to use options, which is a
choice of InAppOptions or PayloadOptions. I went for InAppOptions as
PayloadOptions has an expectation of assets being available which may
not be the case.

[1]: https://github.com/alphagov/whitehall/blob/a94e27e0869ee086dae8278abaa0a5e96a7b306a/app/validators/safe_html_validator.rb
[2]: https://github.com/alphagov/specialist-publisher/blob/e2782bcfb518ac497e978e48320e556675c4940a/app/validators/safe_html_validator.rb